Fix minor issues related to repositories, services, entities

+++ src/main/java/com/company/project/controller/BacklogController.java
@@ -1,13 +1,15 @@
 package com.company.project.controller;
 
 import java.util.UUID;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import com.company.project.service.BacklogService;
 import com.company.project.model.aggregate.Backlog;
 import org.springframework.beans.factory.annotation.Autowired;
 
 @RestController
-@RequestMapping("/backlogs")
+@RequestMapping("/api/backlogs")
 public class BacklogController {
     private final BacklogService backlogService;
 
@@ -17,17 +19,25 @@ public class BacklogController {
     }
 
     @GetMapping("/{id}")
-    public Backlog getBacklog(@PathVariable UUID id) {
-        return backlogService.getBacklogById(id);
+    public ResponseEntity<Backlog> getBacklogById(@PathVariable UUID id) {
+        return backlogService.getBacklogById(id)
+                .map(backlog -> new ResponseEntity<>(backlog, HttpStatus.OK))
+                .orElseGet(() -> new ResponseEntity<>(HttpStatus.NOT_FOUND));
     }
 
     @PostMapping
-    public void createBacklog(@RequestBody Backlog backlog) {
-        backlogService.createBacklog(backlog);
+    public ResponseEntity<Backlog> createBacklog(@RequestBody Backlog backlog) {
+        Backlog createdBacklog = backlogService.createBacklog(backlog);
+        return new ResponseEntity<>(createdBacklog, HttpStatus.CREATED);
     }
 
     @DeleteMapping("/{id}")
-    public void deleteBacklog(@PathVariable UUID id) {
-        backlogService.deleteBacklog(id);
+    public ResponseEntity<Void> deleteBacklog(@PathVariable UUID id) {
+        if (backlogService.getBacklogById(id).isPresent()) {
+            backlogService.deleteBacklog(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 }
